refactor: replace .opencode with .opencode-plugin
