+++ .opencode-plugin/claude-warning.ts
