+++ .opencode-plugin/diff.ts
