+++ .opencode-plugin/index.ts
