+++ .opencode-plugin/install.ts
