+++ .opencode-plugin/logger.ts
