+++ .opencode-plugin/parse-frontmatter.ts
