+++ .opencode-plugin/update-tool.ts
