+++ .opencode-plugin/update.ts
