--- .opencode/.gitignore
@@ -1,4 +1,0 @@
-command/
-.engram-*
-scripts/
-plugins/
