--- .opencode/agents/artifact-smith.md
@@ -1,1 +1,0 @@
-../../agents/engram-artifact-smith.md
