--- .opencode/agents/assessor.md
@@ -1,1 +1,0 @@
-../../agents/engram-assessor.md
