--- .opencode/agents/curriculum-architect.md
@@ -1,1 +1,0 @@
-../../agents/engram-curriculum-architect.md
