--- .opencode/skills/coach/SKILL.md
@@ -1,1 +1,0 @@
-../../../skills/coach/SKILL.md
