--- .opencode/skills/review/SKILL.md
@@ -1,1 +1,0 @@
-../../../skills/review/SKILL.md
